+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## next deploy
 
 - Fix a bug: for project categories, if there is no category : do not show categories panel in show view, do not show categories input field in edit view
+- Feature: add a filter in members list (admin) to show only "not validated" members
 
 ## v6.0.13 2023 August 28