+++ CHANGELOG.md
@@ -6,6 +6,14 @@
 - Fix a bug: unable to update status to paid for latest payment schedule item
 - Fix a bug: unable to generate statistic
 - Feature: add a filter in members list (admin) to show only "not validated" members
+- Concerning statistics:
+- removes age and type column from all statistics tabs (only in web, not in xlsx export file)
+- index:
+  - renames user column header for projects tab and projects xlsx export
+  - adds group name of user for every tab except projects tab
+  - adds status and project users names for projects tab
+- [TODO DEPLOY] `rails db:seed`
+- [TODO DEPLOY] `rails fablab:es:build_stats`
 - [TODO DEPLOY] `rails fablab:maintenance:regenerate_statistics[2014,1]`
 
 ## v6.0.13 2023 August 28